[fix] 모듈화 및 폴더 분리
