+++ src/index.html
@@ -95,7 +95,7 @@
     <!-- cancelAlert 메세지 -->
     <div id="cancelAlert">
       <div id="cancelAlertTop">
-        <p class="display-bold16"></p>
+        <p class="display-midium16"></p>
       </div>
       <div id="cancelAlertBottom">
         <button id="positiveButton" class="available-medium16">예, 해지합니다.</button>
@@ -105,10 +105,7 @@
   </div>
 
   <!-- <script src="./app.js" type="module"></script> -->
-  <script src="./components/date.js" type="module"></script>
-  <script src="./components/topAndViewer.js" type="module"></script>
-  <script src="./components/fieldTab.js" type="module"></script>
-  <script src="./components/autoRollingNews.js" type="module"></script>
+  <script src="./app.js" type="module"></script>
 </body>
 
 </html>
